Add GtkColorProfile which is a high level abstraction of a color profile suitable for color management
Typically this will be an ICC profile loaded from disk, or embedded in a source
image. A color profile will allow you to change the colors of the image either to
correctly render the colors on a given output device, or to deliberatly
change the colors e.g. introducing a sepia effect on photographs.