Add GtkColorEngine which is a singleton to create GtkColorProfile and GtkColorTransform objects
GtkColorEngine is an object that manages the default color management system
(CMS) which may or may not be installed as plugins.
The GtkColorEngine object automatically loads the default CMS the first time
that it is required.

If no CMM is available then a fallback is used, which does no color
correction to the image whatsoever.