Bug 657770 -- Write releaseed memory in gtkdnd-quartz.c
Changes the GtkDragSourceOwner class to keep a copy of the
source GdkDragContext instead of the GtkDragSourceInfo that's attached
to it as qdata. Get the drag_source_info from the context qdata instead.

Since ordinarily the info (and the context, for that matter) are freed
*before* pasteboardChangeOwner gets called, check that the context is
the same as the current _gdk_quartz_drag_source_context and if they
don't match, NULL the member one and return rather than trying to NULL
info members; if info is NULL, return rather than trying to dereference
it. This prevents accessing the already-freed info structure and
context.

Set the info pointer to NULL after clearing its contents and freeing
it (probably unnecessary, but cheap insurance).